requirements.txt
asyncpg==0.29.0
discord.py==2.3.2
numpy==1.26.4
